Add uploadthing route and integration Add uploadthing library and dependencies Update environment variables Fix import in product-drawer component Update page component to handle authentication Update button component styles Update table-card component to use qrCode property Add qrcode action for validation Add upload icon to icons component Update create-qr-button component to refresh router after creating QR code Update env.js to include UPLOADTHING_SECRET and UPLOADTHING_APP_ID Add promotion action to create new promotion

+++ src/lib/actions/table.ts
@@ -2,18 +2,34 @@
 
 import { revalidatePath } from 'next/cache'
 import { db } from '../../server/db'
+import { getServerAuthSession } from '../../server/auth'
 
 export const getTables = async () => {
-  const tables = await db.table.findMany()
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
+  const tables = await db.table.findMany({
+    include: {
+      orders: true,
+    },
+  })
 
   revalidatePath('/dashboard/table')
   return tables
 }
 
 export const getTableById = async (id: number) => {
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
   const table = await db.table.findUnique({
     where: {
       id,
+    },
+    include: {
+      orders: true,
     },
   })
 
@@ -22,6 +38,10 @@ export const getTableById = async (id: number) => {
 }
 
 export const createTable = async (name: string) => {
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
   const table = await db.table.create({
     data: {
       name,
@@ -33,6 +53,10 @@ export const createTable = async (name: string) => {
 }
 
 export const updateTable = async (id: number, name: string) => {
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
   const table = await db.table.update({
     where: {
       id,
@@ -47,6 +71,10 @@ export const updateTable = async (id: number, name: string) => {
 }
 
 export const deleteTable = async (id: number) => {
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
   const table = await db.table.delete({
     where: {
       id,
@@ -58,6 +86,10 @@ export const deleteTable = async (id: number) => {
 }
 
 export const addQrCode = async (id: number, qrCode: string) => {
+  const session = await getServerAuthSession()
+
+  if (!session) throw new Error('Unauthorized')
+
   const table = await db.table.update({
     where: {
       id,
